+++ KillSystem/src/main/java/com/KillSystem/controller/ShippingAddressController.java
@@ -14,13 +14,22 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.KillSystem.Service.ShippingAddressService;
+import com.KillSystem.common.ResponseCode;
 import com.KillSystem.common.ServerResponse;
 import com.KillSystem.domain.Goods;
-import com.KillSystem.domain.Order;
 import com.KillSystem.domain.ShippingAddress;
 import com.KillSystem.domain.User;
 import com.github.pagehelper.PageInfo;
 
+/**
+ * @author xcxcxcxcx
+ * 
+ * 收货地址controller
+ * 包括收货地址的增删查
+ * 
+ * 2018年4月5日
+ *
+ */
 @Controller
 public class ShippingAddressController {
 	
@@ -29,6 +38,9 @@ public class ShippingAddressController {
 
 	@RequestMapping("/getShippingAddress.do")
 	public ModelAndView getShippingAddress(Goods goods,HttpServletRequest request,HttpSession session) {
+		if (session.getAttribute("tel_num") == null || session.getAttribute("passwd") == null) {
+			return new ModelAndView("fail", null);
+		}
 		User user = new User();
 		user.setTel_num((String)session.getAttribute("tel_num"));
 		request.setAttribute("goods_name", goods.getGoods_name());
@@ -42,6 +54,10 @@ public class ShippingAddressController {
 	@RequestMapping("/insertShippingAddress.do")
 	@ResponseBody
 	public ServerResponse insertShippingAddress(ShippingAddress shippingAddress,HttpServletRequest request,HttpSession session,@RequestParam(value = "upload_file",required = false) MultipartFile file) {
+		if (session.getAttribute("tel_num") == null || session.getAttribute("passwd") == null) {
+			return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(), "用户未登录");
+		}
+		
 		if(shippingAddressService.insert(shippingAddress) == 1) {
 			return ServerResponse.createBySuccessMessage("新增收货地址成功!");
 		}
@@ -52,6 +68,10 @@ public class ShippingAddressController {
 	@RequestMapping("/deleteShippingAddress.do")
 	@ResponseBody
 	public ServerResponse deleteShippingAddress(ShippingAddress shippingAddress,HttpServletRequest request,HttpSession session,@RequestParam(value = "upload_file",required = false) MultipartFile file) {
+		if (session.getAttribute("tel_num") == null || session.getAttribute("passwd") == null) {
+			return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(), "用户未登录");
+		}
+		
 		if(shippingAddressService.delete(shippingAddress) == 1) {
 			return ServerResponse.createBySuccessMessage("删除收货地址成功!");
 		}
